+++ src/backends/preload_elf/aarch64/arch_spec_bottom.c
@@ -115,7 +115,7 @@ static unsigned long wrapper_frame_get_arg(
 	}
 }
 
-long retrace_as_call_real(const void *real_impl,
+intptr_t retrace_as_call_real(const void *real_impl,
 	const struct FuncParam params[],
 	int params_cnt)
 {
@@ -179,7 +179,7 @@ int retrace_as_setup_params(
 		 * anything beyond goes on the stack starting at orig_sp
 		 */
 		params[param_idx].val =
-			(long) wrapper_frame_get_arg(frame, param_idx);
+			(intptr_t) wrapper_frame_get_arg(frame, param_idx);
 	}
 
 	/* set up varargs params */
@@ -247,7 +247,7 @@ int retrace_as_setup_params(
 		params[param_idx].data_type = dt;
 
 		params[param_idx].val =
-			(long) wrapper_frame_get_arg(frame, param_idx);
+			(intptr_t) wrapper_frame_get_arg(frame, param_idx);
 	}
 
 	*params_cnt = param_idx;
@@ -268,7 +268,7 @@ void retrace_as_cancel_sched_real(void *arch_spec_ctx)
 	((struct WrapperAArch64Frame *) arch_spec_ctx)->call_real_flag = 0;
 }
 
-void retrace_as_set_ret_val(void *arch_spec_ctx, long ret_val)
+void retrace_as_set_ret_val(void *arch_spec_ctx, intptr_t ret_val)
 {
 	struct WrapperAArch64Frame *frame = arch_spec_ctx;
